% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Output.R
\name{output_format}
\alias{output_format}
\alias{out_file}
\alias{out_memory}
\alias{out_nothing}
\title{Specifying Output Format}
\usage{
out_file(path, ...)

out_memory(...)

out_nothing(...)
}
\arguments{
\item{path}{Path to the output file.}

\item{\dots}{Currently unused.}
}
\value{
An object of class "Output" as required by
\code{$\link{output}()} and \code{$\link{predict}()}.
}
\description{
Functions in this page are used to specify the format of output results.
They are intended to provide the argument of functions such as
\code{$\link{output}()} and \code{$\link{predict}()}.
Currently there are three types of output: \code{out_file()} indicates
that the result should be written into a file, \code{out_memory()} makes
the result to be returned as R objects, and \code{out_nothing()} means
the result is not needed and will not be returned.
}
\seealso{
\code{$\link{output}()}, \code{$\link{predict}()}
}
\author{
Yixuan Qiu <\url{https://statr.me}>
}
