% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Reco}
\alias{Reco}
\title{Construct a recommender system object}
\usage{
Reco()
}
\value{
\code{Reco()} returns an object of class "\code{RecoSys}"
equipped with methods
\code{$\link{convert_train}()}, \code{$\link{convert_train}()},
\code{$\link{train}()} and \code{$\link{predict}()}, which describe
the typical process of reading data, building model and
predicting results. See their help documents for details.
}
\description{
This function simply returns an object of class "\code{RecoSys}"
that can be used to construct recommender model and conduct prediction.
}
\author{
Yixuan Qiu <\url{http://statr.me}>
}
\references{
LIBMF: A Matrix-factorization Library for Recommender Systems.
\url{http://www.csie.ntu.edu.tw/~cjlin/libmf/}

Y. Zhuang, W.-S. Chin, Y.-C. Juan, and C.-J. Lin.
A Fast Parallel Stochastic Gradient Method for Matrix Factorization in Shared Memory Systems.
Technical report 2014.
}
\seealso{
\code{\link{convert}}, \code{\link{train}}, \code{\link{predict}}
}
\keyword{models}

