% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RecoSys.R
\name{Reco}
\alias{Reco}
\title{Constructing a Recommender System Object}
\usage{
Reco()
}
\value{
\code{Reco()} returns an object of class "\code{RecoSys}"
equipped with methods
\code{$\link{tune}()}, \code{$\link{train}()}, \code{$\link{output}()}
and \code{$\link{predict}()}, which describe the typical process of
building and tuning model, outputing coefficients, and
predicting results. See their help documents for details.
}
\description{
This function simply returns an object of class "\code{RecoSys}"
that can be used to construct recommender model and conduct prediction.
}
\author{
Yixuan Qiu <\url{http://statr.me}>
}
\references{
W.-S. Chin, Y. Zhuang, Y.-C. Juan, and C.-J. Lin.
A Fast Parallel Stochastic Gradient Method for Matrix Factorization in Shared Memory Systems.
ACM TIST, 2015.

W.-S. Chin, Y. Zhuang, Y.-C. Juan, and C.-J. Lin.
A learning-rate schedule for stochastic gradient methods to matrix factorization.
PAKDD, 2015.
}
\seealso{
\code{$\link{tune}()}, \code{$\link{train}()}, \code{$\link{output}()},
\code{$\link{predict}()}
}
\keyword{models}

